+++ test/fixtures/ignore-unused-exports-comments/a.mjs
@@ -1,0 +1,3 @@
+// ignore unused exports
+export const a = true;
+export default true;