+++ test/fixtures/ignore-unused-exports-comments/b.mjs
@@ -1,0 +1,3 @@
+// ignore unused exports default
+export const a = true;
+export default true;