+++ test/fixtures/ignore-unused-exports-comments/c.mjs
@@ -1,0 +1,4 @@
+// ignore unused exports a, b
+export const a = true;
+export const b = true;
+export default true;